# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apple Connector API",
    "description": "Hybrid HTTP API for Messages.app, Reminders.app, Notes.app, Calendar.app, and Contacts.\n\n**Reads** use live SQLite connections (requires Full Disk Access).\n\n**Writes** for Reminders, Calendar events, and Contacts use EventKit / Contacts framework (requires Reminders, Calendars, and Contacts permissions in System Settings). Unsupported fields return `422`; smart lists, read-only calendars/containers return `403`. Post-save responses hydrate from SQLite with up to 5×100ms retries; `sync_pending: true` indicates the SQLite read path has not caught up yet.\n\nPagination uses keyset cursors only (default limit 50, maximum 200, newest first). Offsets are not supported.\n\nAuthentication, TLS, and network exposure controls are expected to be enforced by an external reverse proxy or firewall. This service does not implement authentication or TLS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this service (use the host you opened in the browser, e.g. http://127.0.0.1:3000 or http://localhost:3000)"
    }
  ],
  "paths": {
...
    "/v1/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "List contact groups",
        "operationId": "listGroups",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{group_id}": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get a contact group",
        "operationId": "getGroup",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryDto"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "groups"
        ],
        "summary": "Delete a group",
        "operationId": "deleteGroup",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Group deleted"
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases or Contacts framework unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "groups"
        ],
        "summary": "Update a group",
        "operationId": "updateGroup",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSummaryDto"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases or Contacts framework unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{group_id}/contacts": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts in a group as JSON",
        "operationId": "listGroupContacts",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated contacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{group_id}/contacts/carddav": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts in a group as CardDAV XML",
        "operationId": "listGroupContactsCarddav",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "CardDAV multistatus",
            "content": {
              "application/carddav+xml": {}
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{group_id}/contacts/vcard": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts in a group as vCard",
        "operationId": "listGroupContactsVcard",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "vCard feed",
            "content": {
              "text/vcard": {}
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactId": {
        "type": "string",
        "description": "Stable identifier for a contact (its UUID).",
        "example": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
      },
...
      "ContactPageDto": {
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSummaryDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        }
      },
...
      "ContactSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "source_id",
          "container_id"
        ],
        "properties": {
          "container_id": {
            "$ref": "#/components/schemas/ContainerId"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/ContactId"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "modification_date": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp"
              }
            ]
          },
          "organization": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_id": {
            "$ref": "#/components/schemas/SourceId"
          }
        }
      },
...
      "ContainerId": {
        "type": "string",
        "description": "Stable identifier for a contact container (its UUID).",
        "example": "11111111-1111-1111-1111-111111111111"
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_error",
          "not_found",
          "range_not_satisfiable",
          "service_unavailable",
          "forbidden",
          "conflict",
          "internal_error"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "GroupId": {
        "type": "string",
        "description": "Stable identifier for a contact group (its UUID).",
        "example": "22222222-2222-2222-2222-222222222222"
      },
...
      "GroupPageDto": {
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupSummaryDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        }
      },
      "GroupSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "source_id",
          "container_id",
          "is_smart_group",
          "is_subscribed"
        ],
        "properties": {
          "container_id": {
            "$ref": "#/components/schemas/ContainerId"
          },
          "id": {
            "$ref": "#/components/schemas/GroupId"
          },
          "is_smart_group": {
            "type": "boolean"
          },
          "is_subscribed": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_id": {
            "$ref": "#/components/schemas/SourceId"
          }
        }
      },
...
      "PageMetaDto": {
        "type": "object",
        "required": [
          "limit",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean",
            "description": "Whether additional items exist beyond this page."
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "description": "Applied page size for this response.",
            "example": 50,
            "maximum": 200,
            "minimum": 1
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL-safe versioned cursor for the next page when `has_more` is true.",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        }
      },
...
      "SourceId": {
        "type": "string",
        "description": "AddressBook source UUID (directory name under Sources/).",
        "example": "27fd6c1e-5da5-4340-a31a-1d83c25d3b70"
      },
...
      "UnixTimestamp": {
        "type": "integer",
        "format": "int64",
        "description": "Whole seconds since the Unix epoch (`1970-01-01T00:00:00Z`), in UTC.\n\nSerialized as a JSON integer. This replaces the RFC 3339 timestamp strings\nemitted by earlier revisions of the API (see the `apple_types` breaking\nchange in the README).",
        "example": 1705320000
      },
...
      "UpdateGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "reverse_proxy": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication is expected to be enforced by an external reverse proxy or firewall. This API does not implement authentication or TLS."
      }
    }
  },
  "security": [
    {
      "reverse_proxy": []
    }
  ],
...
}